refactor(indent): extract utils with zero logic change
